+++ docs/changelog.md
@@ -18,7 +18,8 @@
 
 ### Added
 
-- Command `cbd aws list-roles` will lisst all role names
+- Command `remote-cloudbreak-shell` will show can use cloudbreak shell on local machine
+- Command `cbd aws list-roles` will list all role names
 - Command `cbd azure configure-arm` will create your arm application which can used by cloudbreak
 - Command `cbd azure deploy-dash` will deploy a dash application in your Azure account
 - Command `cbd start` will execute the migration by default. If SKIP_DB_MIGRATION_ON_START envvar set to true in Profile, the migration will be skipped